operator/multi-pool: Rework multi-pool allocator files structure
Move the multi-pool allocator into a package under
operator/pkg/ipam/allocator. Besides mimicking the same structure that
we already have for the other IPAM allocators, this allows to streamline
the multi-pool cell file gated by the "ipam_provider_operator" build
tag. Having a cell.go with the build tag is still essential to exclude
the multi-pool allocator implementation from the operator binary when
not requested, but at the same time the new multipool package can be
referenced in testing without adding any build tag.

A script based test validating the multi-pool allocator module will be
added in a subsequent commit.
